+++ Libs/DeathNotificationLib/Locales/strings_de.lua
@@ -1,7 +1,7 @@
 local _dnl = DeathNotificationLib.Internal ---@class _dnl
 if not _dnl then return end
 _dnl.L.de.STRINGS = {
-    death_alert_font = "Fonts\\blei00d.TTF",
+    death_alert_font = "Fonts\\FRIZQT__.TTF",
     death_alert_default_message = "<name>( der <race> <class>)[ von <guild>] wurde\nvon <source>[ auf Stufe <level>][ in <zone>] getötet.",
     death_alert_default_fall_message = "<name>( der <race> <class>) ist\n[auf Stufe <level> ][in <zone> ]zu Tode gestürzt.",
     death_alert_default_drown_message = "<name>( der <race> <class>) ist\n[auf Stufe <level> ][in <zone> ]ertrunken.",