+++ Libs/DeathNotificationLib/Locales/strings_es.lua
@@ -1,7 +1,7 @@
 local _dnl = DeathNotificationLib.Internal ---@class _dnl
 if not _dnl then return end
 _dnl.L.es.STRINGS = {
-    death_alert_font = "Fonts\\blei00d.TTF",
+    death_alert_font = "Fonts\\FRIZQT__.TTF",
     death_alert_default_message = "<name>( <race> <class>)[ de <guild>] ha muerto\na manos de <source>[ en nivel <level>][ en <zone>].",
     death_alert_default_fall_message = "<name>( <race> <class>) ha muerto\nde una caída[ en nivel <level>][ en <zone>].",
     death_alert_default_drown_message = "<name>( <race> <class>) se ha ahogado\n[ en nivel <level>][ en <zone>].",