+++ Libs/DeathNotificationLib/Locales/strings_fr.lua
@@ -1,7 +1,7 @@
 local _dnl = DeathNotificationLib.Internal ---@class _dnl
 if not _dnl then return end
 _dnl.L.fr.STRINGS = {
-    death_alert_font = "Fonts\\blei00d.TTF",
+    death_alert_font = "Fonts\\FRIZQT__.TTF",
     death_alert_default_message = "<name>( <race> <class>)[ de <guild>] a été tué\npar <source>[ au niveau <level>][ à <zone>].",
     death_alert_default_fall_message = "<name>( <race> <class>) est tombé\nà mort[ au niveau <level>][ à <zone>].",
     death_alert_default_drown_message = "<name>( <race> <class>) s'est noyé\n[ au niveau <level>][ à <zone>].",