+++ Libs/DeathNotificationLib/Locales/strings_pt.lua
@@ -1,7 +1,7 @@
 local _dnl = DeathNotificationLib.Internal ---@class _dnl
 if not _dnl then return end
 _dnl.L.pt.STRINGS = {
-    death_alert_font = "Fonts\\blei00d.TTF",
+    death_alert_font = "Fonts\\FRIZQT__.TTF",
     death_alert_default_message = "<name>( <race> <class>)[ de <guild>] foi morto\npor <source>[ no nível <level>][ em <zone>].",
     death_alert_default_fall_message = "<name>( <race> <class>) caiu\naté a morte[ no nível <level>][ em <zone>].",
     death_alert_default_drown_message = "<name>( <race> <class>) se afogou\n[ no nível <level>][ em <zone>].",